Adding Rubocop + Fixing some of its comments
Adding some custom config:
- Layout/EmptyLinesAroundClassBody: excluding Job files. We leave an
empty line between the class declaration and the job trigger to make it
easier to read
- Metrics/BlockLength: excluding specs to ignore the length of
RSpec `describe` and `context` blocks
- Metrics/LineLength: setting it to 100 instead of 80
- Metrics/MethodLength: setting it to 20. 10 lines can be somewhat
acceptable, but 20 is definitely a code smell
- Style/Documentation: The code doesn't have any comments and getting
that report is generating a lot of noise. We'll try to introduce the
documentation progressively, and at that point we may want to enable the
rule again
- Style/FrozenStringLiteralComment: analyzing only Ruby file (e.g.
excluding config, Gemfile, Rakefile)

+++ spec/support/dynamoid.rb
@@ -4,9 +4,7 @@ module DynamoidReset
   def self.all
     Dynamoid.adapter.list_tables.each do |table|
       # Only delete tables in our namespace
-      if table =~ /^#{Dynamoid::Config.namespace}/
-        Dynamoid.adapter.delete_table(table)
-      end
+      Dynamoid.adapter.delete_table(table) if table =~ /^#{Dynamoid::Config.namespace}/
     end
     Dynamoid.adapter.tables.clear
     # Recreate all tables to avoid unexpected errors
